Finally lists are working again
Header has a more consistent behaviour
Both Feed and Browse are using more consistent information
of books, so the problem with different books click should be gone.

Search update based on loading flag is triggered always, so the
problem with corrupted/unsynched data is gone.
Abstracted some methods from feed and browse, so both should use the
same behaviour.
Search is shown on main page, and works correctly with the
search.

Removed the customRouter implementation, now we can achieve
most of the functionality with the provided drawer navigator.
Refactored some navigation code to accomodate this.

TODO: Abstract loader. it's getting huge.